Jinja populating dashboard
Added jinja logic to fill in dashboard videos in a series. Uses for loop
and a dictionary of videos in context

Also added query of videos to views. Right now it populates it by videos
you uploaded, later will be changed to videos in a series, but the
concept is implemented