Added course/courseloader pages
These pages are going to be similar to the video pages, and i have used
them as a base for now. The goal will be to change these pages with
additions for courses, such as next./previous buttons. This should be
simple because in the views we can just increment the position of the
course to change the dispalyed page